Extract calendly link into component
